Added engines, ACTION BUTTONS NOW WORK!
Moved thrusters to spacecraft.dm, added engines as a new stock parts tree.  They are currently apart of the same research nodes as manipulators, in a non-intrusive way. THE ACTION BUTTONS ARE NOW FUNCTIONAL! That took way too freaking long to figure out, especially how to turn on the god damn lights. It works, now.